ChatBot/requirements.txt
Flask
psycopg2-binaryv